requirements.txt
# 기본 예시 — 필요에 맞게 수정하세요
# PyTorch CUDA 휠 인덱스
--extra-index-url https://download.pytorch.org/whl/cu121

# Core DL (CUDA)
torch==2.4.0
torchvision==0.19.0

# Utils / Numerics
numpy==1.23.5
Pillow==10.4.0
pandas==2.2.2

# Embedding / Models
open_clip_torch==2.26.1
timm==1.0.9
transformers==4.45.2
tokenizers==0.20.0
safetensors==0.4.5

# Vector Search (GPU)
faiss-gpu-cu12==1.12.0

notebook==6.5.4
ipykernel==6.28.0
